requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
httpx==0.27.0
pydantic==2.7.1
google-cloud-bigquery==3.24.0
google-cloud-aiplatform==1.57.0
vertexai==1.57.0
python-dotenv==1.0.1
websockets==12.0
reportlab==4.2.0
google-api-python-client==2.131.0
 